Minimal (but still violent) change to build on modern GHCs.
The HOC package has been split in 2; one for the objective-c bits and one for the rest.  The old hacks for building ObjC code are no longer necessary, and splitting the package makes it so we don't need the hack for loading the ObjC code during compiling (for Template Haskell use).

I plan to clean this up a bit more too; I'll probably make a "HOC.CBits.*" module for each obj-c file containing nothing but the foreign imports, so the haskell types are declared in the same package as the Obj-C code being imported.