New gnustep-config package
This provides a simple interface to the gnustep-config command-line
tool.  Unfortunately I don't think we can easily avoid depending on it
on the Mac because it's used in HOC-cbits' Setup.hs, and we can't use an
"if" on the Build-Type field.

Simplified that Setup.hs by extracting all of the gnustep-config logic,
also used it in a couple other places where paths were hard-coded.