Added an NSRange-return test
This tests the returning of multiple 64-bit integers.  I feel this is
useful in addition to the NSRect test because that one passes multiple
floating-point values, which, if passed in registers, are passed
differently than ints.

Unfortunately I think in both cases they are passed in memory due to
llvm-gcc's wrong ABI implementation (maybe GCC's too since the test
passes on linux with the same cifIsStret function), making
the issue moot.