Switched to new HSO implementation
The new implementation is in the objc-ffi package.

A fair amount of testing is still needed, and the allocation stats
system needs some work (there's a simplistic surrogate in place but it
doesn't work because it tries to install many finalizers on the same
HSO, and the Haskell runtime drops some).

The test suite fails one of the tests due to the miscounting in the
stats, but I have traced that case through by hand and I believe the
system is actually operating correctly in that case.